Create README.md
Price prediction uses an algorithm to analyze a product or service based on its characteristics, demand, and current market trends. Then the software sets a price at a level it predicts will both attract customers and maximize sales. In some circles, the practice is called price forecasting or predictive pricing.